build(npm): streamlining peer deps even more
