+++ .npmrc
@@ -1,2 +1,1 @@
 strict-peer-dependencies=false
-auto-install-peers=true
