--- helpers/enzyme-redux/index.d.ts
@@ -1,11 +1,0 @@
-import {ReactWrapper, ShallowWrapper} from 'enzyme';
-import {ReactElement} from 'react';
-
-export function shallowWithStore<P>(
-    Component: ReactElement<P>,
-    store: any,
-    withContext: boolean = true
-): ShallowWrapper<P>;
-export function mountWithStore<P>(Component: ReactElement<P>, store: any, withContext: boolean = true): ReactWrapper<P>;
-export function shallowWithState<P>(Component: ReactElement<P>, state: any): ShallowWrapper<P>;
-export function mountWithState<P>(Component: ReactElement<P>, state: any): ReactWrapper<P>;
