--- helpers/enzyme-redux/package.json
@@ -1,17 +1,0 @@
-{
-    "name": "@helpers/enzyme-redux",
-    "version": "41.3.0",
-    "private": true,
-    "description": "",
-    "keywords": [],
-    "license": "ISC",
-    "author": "",
-    "main": "index.jsx",
-    "types": "index.d.ts",
-    "dependencies": {
-        "enzyme": "3.11.0",
-        "react": "17.0.2",
-        "react-redux": "7.2.6",
-        "tslib": "1.14.1"
-    }
-}
