+++ packages/react/package.json
@@ -85,11 +85,9 @@
         "react-dnd-html5-backend": "15.1.3",
         "react-infinite-scroll-component": "4.5.3",
         "react-modal": "3.15.1",
-        "react-redux": "7.2.6",
         "react-tether": "1.0.5",
         "react-textarea-autosize": "8.3.3",
         "react-transition-group": "2.9.0",
-        "redux": "4.2.0",
         "reselect": "4.1.5",
         "split-on-first": "3.0.0",
         "strict-uri-encode": "2.0.0",
@@ -97,7 +95,6 @@
         "underscore.string": "3.3.6"
     },
     "devDependencies": {
-        "@helpers/enzyme-redux": "workspace:*",
         "@hot-loader/react-dom": "17.0.2",
         "@swc/cli": "0.1.57",
         "@swc/core": "1.2.174",
@@ -124,7 +121,7 @@
         "@types/react-infinite-scroll-component": "4.2.5",
         "@types/react-modal": "3.13.1",
         "@types/react-redux": "7.1.22",
-        "@types/react-tether": "0.5.5",
+        "@types/react-tether": "1.0.0",
         "@types/react-textarea-autosize": "4.3.6",
         "@types/react-transition-group": "2.9.2",
         "@types/redux-logger": "3.0.9",
@@ -157,6 +154,8 @@
         "react": "17.0.2",
         "react-dnd-test-backend": "15.1.2",
         "react-dom": "17.0.2",
+        "react-redux": "7.2.6",
+        "redux": "4.2.0",
         "redux-logger": "3.0.6",
         "redux-mock-store": "1.5.4",
         "redux-promise-middleware": "6.1.2",
@@ -169,7 +168,11 @@
         "typescript": "4.6.4"
     },
     "peerDependencies": {
-        "react": "17.x.x",
-        "react-dom": "^16.14.0"
+        "@types/react": "^17.0",
+        "@types/react-dom": "^17.0",
+        "react": "^17.0",
+        "react-dom": "^17.0",
+        "react-redux": "^7.2.6",
+        "redux": "^4"
     }
 }
