+++ packages/react/src/components/blankSlate/tests/BlankSlateWithTable.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from '@helpers/enzyme-redux';
+import {shallowWithState} from '@test-utils';
 
 import {BlankSlateWithTable} from '../BlankSlatesHOC';
