+++ packages/react/src/components/datePicker/tests/DatePickerDropdown.spec.tsx
@@ -1,5 +1,5 @@
 import {shallow, ShallowWrapper} from 'enzyme';
-import {shallowWithState} from '@helpers/enzyme-redux';
+import {shallowWithState} from '@test-utils';
 import moment from 'moment';
 import * as _ from 'underscore';
 import {render, screen} from '@test-utils';
