+++ packages/react/src/components/flippable/tests/Flippable.spec.tsx
@@ -1,5 +1,5 @@
 import {mount, ReactWrapper, shallow} from 'enzyme';
-import {shallowWithState} from '@helpers/enzyme-redux';
+import {shallowWithState} from '@test-utils';
 
 import {BrowserUtils} from '../../../utils/BrowserUtils';
 import {removeTestAppContainer} from '../../../utils/tests/TestUtils';
