+++ packages/react/src/components/modal/tests/ModalCompositeConnected.spec.tsx
@@ -1,5 +1,5 @@
 import {ShallowWrapper} from 'enzyme';
-import {shallowWithState, shallowWithStore} from '@helpers/enzyme-redux';
+import {shallowWithState, shallowWithStore} from '@test-utils';
 import ReactModal from 'react-modal';
 
 import {getStoreMock} from '../../../utils/tests/TestUtils';
