+++ packages/react/src/components/multilineBox/tests/MultilineBox.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState, shallowWithStore} from '@helpers/enzyme-redux';
+import {shallowWithState, shallowWithStore} from '@test-utils';
 import * as _ from 'underscore';
 
 import {StringListActions} from '../../../reusableState/customList/StringListActions';
