+++ packages/react/src/components/numericInput/tests/NumericInputConnected.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithStore} from '@helpers/enzyme-redux';
+import {shallowWithStore} from '@test-utils';
 import {render, screen} from '@test-utils';
 import {ShallowWrapper} from 'enzyme';
