+++ packages/react/src/components/pagination/tests/PaginationPagesNumber.spec.tsx
@@ -1,5 +1,5 @@
 import {ShallowWrapper} from 'enzyme';
-import {shallowWithState, shallowWithStore} from '@helpers/enzyme-redux';
+import {shallowWithState, shallowWithStore} from '@test-utils';
 
 import {getStoreMock} from '../../../utils/tests/TestUtils';
 import {PaginationReduxActions} from '../../navigation/pagination';
