+++ packages/react/src/components/table-hoc/tests/TableWithBlankslate.spec.tsx
@@ -1,5 +1,5 @@
 import {shallow} from 'enzyme';
-import {shallowWithState} from '@helpers/enzyme-redux';
+import {shallowWithState} from '@test-utils';
 import * as _ from 'underscore';
 
 import {BlankSlateWithTable} from '../../blankSlate';
