+++ packages/react/src/components/table-hoc/tests/TableWithDatePicker.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithState} from '@helpers/enzyme-redux';
+import {shallowWithState} from '@test-utils';
 import moment from 'moment';
 import * as _ from 'underscore';
