+++ packages/react/src/components/table-hoc/tests/TableWithPagination.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithStore} from '@helpers/enzyme-redux';
+import {shallowWithStore} from '@test-utils';
 import {ComponentType} from 'react';
 import * as _ from 'underscore';
