+++ packages/react/src/components/validation/hoc/tests/WithDirtyStickyFooter.spec.tsx
@@ -1,4 +1,4 @@
-import {shallowWithStore} from '@helpers/enzyme-redux';
+import {shallowWithStore} from '@test-utils';
 
 import {getStoreMock} from '../../../../utils/tests/TestUtils';
 import {StickyFooter} from '../../../stickyFooter/StickyFooter';
