+++ packages/react/src/hoc/withPreventNavigation/tests/modalWithPreventNavigation.spec.tsx
@@ -1,5 +1,5 @@
 import {ShallowWrapper} from 'enzyme';
-import {shallowWithStore} from '@helpers/enzyme-redux';
+import {shallowWithStore} from '@test-utils';
 import {Component} from 'react';
 
 import {closeModal} from '../../../components/modal/ModalActions';
