+++ packages/website/package.json
@@ -13,7 +13,7 @@
     "main": "src/Index.ts",
     "scripts": {
         "build": "next build && next export && next-sitemap",
-        "clean": "rimraf dist built",
+        "clean": "rimraf dist built out .next",
         "lintfix": "../../node_modules/.bin/stylelint --fix --syntax \"**/*.scss\" && ../../node_modules/.bin/prettier --write \"**/*.{scss,ts,tsx,js,jsx,json,md,yml,html}\" && ../../node_modules/.bin/eslint -c ./.eslintrc.js \"src/**/*.{ts,tsx}\" --fix",
         "precommit": "lint-staged",
         "prod:server": "next start",
@@ -49,11 +49,8 @@
         "@typescript/twoslash": "3.1.0",
         "@typescript/vfs": "1.3.5",
         "classnames": "2.3.1",
-        "codemirror": "5.63.3",
-        "d3": "3.5.17",
         "faker": "4.1.0",
         "flagg": "1.1.2",
-        "jquery": "3.6.0",
         "lorem-ipsum": "2.0.4",
         "lz-string": "1.4.4",
         "moment": "2.29.4",
@@ -71,16 +68,13 @@
         "react-dom": "17.0.2",
         "react-markdown": "7.1.2",
         "react-redux": "7.2.6",
-        "react-router": "6.3.0",
-        "react-router-dom": "6.3.0",
         "redux": "4.2.0",
         "redux-devtools-extension": "2.13.9",
         "redux-promise-middleware": "6.1.2",
         "redux-thunk": "2.4.1",
         "remark-gfm": "3.0.1",
         "reselect": "4.1.5",
-        "underscore": "1.13.1",
-        "underscore.string": "3.3.6"
+        "underscore": "1.13.1"
     },
     "devDependencies": {
         "@types/classnames": "2.3.0",
@@ -94,8 +88,6 @@
         "@types/react-dom": "17.0.15",
         "@types/react-modal": "3.13.1",
         "@types/react-redux": "7.1.22",
-        "@types/react-router": "5.1.18",
-        "@types/react-router-dom": "5.3.3",
         "@types/redux-promise-middleware": "6.0.0",
         "@types/underscore": "1.8.10",
         "@types/webpack-env": "1.16.4",
